pyproject.toml
[tool.poetry]
name = "djang-sample-project"
version = "0.1.0"
description = "Sample project for Python Django"
authors = ["Kuriz15 <chrisbutalid@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
djangorestframework = "^3.14.0"
whitenoise = {extras = ["brotli"], version = "^6.5.0"}
ipython = "^8.14.0"
psycopg = "^3.1.10"
django-extensions = "^3.2.3"
django-structlog = "^5.3.0"
httpx = "^0.24.1"
drf-spectacular = {extras = ["sidecar"], version = "^0.26.4"}
django-environ = "^0.10.0"
dj-rest-auth = {extras = ["with-social"], version = "^4.0.1"}
djangorestframework-dataclasses = "^1.2.0"
django-filter = "^23.2"
django-debug-toolbar = "^4.2.0"
psycopg2 = "^2.9.9"
djoser = "^2.2.2"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
isort = "^5.12.0"
pytest = "^7.4.0"
pre-commit = "^3.3.3"
flake8-isort = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
